[CodeBuild] changes from code review
