+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "exasol-transformers-extension"
-version = "0.6.0"
+version = "0.7.0"
 description = "An Exasol extension to use state-of-the-art pretrained machine learning models via the transformers api."
 
 authors = [
